数据库连接配置修改

+++ youlai-service/src/main/java/com/youlai/service/system/mapper/SysPermissionMapper.java
@@ -1,16 +1,14 @@
 package com.youlai.service.system.mapper;
 
-
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 import com.youlai.service.system.entity.SysPermission;
-import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Select;
+import org.mapstruct.Mapper;
 
 import java.util.List;
 
 @Mapper
 public interface SysPermissionMapper extends BaseMapper<SysPermission> {
-
     @Select("SELECT e.permission_id,e.permission_name, e.perms  " +
             "    FROM  sys_user a " +
             "    INNER JOIN sys_user_role b ON b.user_id = a.user_id " +
@@ -18,5 +16,7 @@ public interface SysPermissionMapper extends BaseMapper<SysPermission> {
             "    INNER JOIN sys_role_permission d ON d.role_id = c.role_id " +
             "    INNER JOIN sys_permission e ON e.permission_id = d.permission_id " +
             "    WHERE a.user_id =#{userId}")
-    List<SysPermission> listByUserId(Long userId);
+    List<SysPermission> listByUserId(Integer userId);
+
+
 }
